requirements.txt
# FILE: requirements.txt
psycopg2-binary
mysql-connector-python
oracledb
pyodbc
msal
requests
schedule
python-dotenv
ollama